--- Assets/Scripts/MenuManager.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class MenuManager : MonoBehaviour {
-	public void LoadScene(string name){
-		SceneManager.LoadScene(name);
-	}
-
-	public void QuitGame(){
-		Application.Quit();
-	}
-}
